# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YouBike 2.0 Taipei Station API",
    "description": "\n以 **臺北市 YouBike 2.0 站點即時資訊** Open Data 為主題的 RESTful API。\n\n資料來源：臺北市政府交通局（data.taipei）\n`https://tcgbusfs.blob.core.windows.net/dotapp/youbike/v2/youbike_immediate.json`\n\n## Resources\n| Resource | 說明 |\n|---|---|\n| **Areas** | 行政區（由站點資料的 `sarea` 欄位整理而成） |\n| **Stations** | YouBike 2.0 站點：位置、車位、即時可借 / 可還數量 |\n| **Stats & Admin** | 車況統計與 Open Data 重新匯入 |\n\n所有端點皆以 `/api/v1` 為前綴。資料儲存於 SQLite，首次啟動若資料庫為空會自動匯入 `data/youbike_stations.json`。\n",
    "contact": {
      "name": "IoTApp Open Data RESTful API Homework"
    },
    "license": {
      "name": "政府資料開放授權條款－第1版",
      "url": "https://data.gov.tw/license"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/areas/{area_id}/stations": {
      "get": {
        "tags": [
          "Areas"
        ],
        "summary": "列出行政區內的站點",
        "operationId": "list_area_stations_api_v1_areas__area_id__stations_get",
        "parameters": [
          {
            "name": "area_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Area Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_StationOut_"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stations": {
      "get": {
        "tags": [
          "Stations"
        ],
        "summary": "搜尋 / 列出站點（分頁）",
        "operationId": "list_stations_api_v1_stations_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "關鍵字（站名、英文站名、地址模糊比對）",
              "examples": [
                "捷運"
              ],
              "title": "Q"
            },
            "description": "關鍵字（站名、英文站名、地址模糊比對）"
          },
          {
            "name": "area_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "行政區 id",
              "title": "Area Id"
            },
            "description": "行政區 id"
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "是否營運中",
              "title": "Is Active"
            },
            "description": "是否營運中"
          },
          {
            "name": "min_bikes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "可借車輛數下限",
              "title": "Min Bikes"
            },
            "description": "可借車輛數下限"
          },
          {
            "name": "min_docks",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "可還空位數下限",
              "title": "Min Docks"
            },
            "description": "可還空位數下限"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "sno",
                "name_zh",
                "available_bikes",
                "available_docks",
                "total_docks",
                "updated_at"
              ],
              "type": "string",
              "default": "sno",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "asc",
              "title": "Order"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_StationOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Stations"
        ],
        "summary": "新增站點",
        "operationId": "create_station_api_v1_stations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationOut"
                }
              }
            }
          },
          "409": {
            "description": "sno 重複",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/stations/nearby": {
      "get": {
        "tags": [
          "Stations"
        ],
        "summary": "查詢座標附近的站點",
        "operationId": "nearby_api_v1_stations_nearby_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "examples": [
                25.0418
              ],
              "title": "Lat"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "examples": [
                121.5436
              ],
              "title": "Lng"
            }
          },
          {
            "name": "radius",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 5000,
              "exclusiveMinimum": 0,
              "description": "半徑（公尺）",
              "default": 500,
              "title": "Radius"
            },
            "description": "半徑（公尺）"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "only_with_bikes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "只回傳目前有車可借的營運站點",
              "default": false,
              "title": "Only With Bikes"
            },
            "description": "只回傳目前有車可借的營運站點"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StationNearbyOut"
                  },
                  "title": "Response Nearby Api V1 Stations Nearby Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stations/{sno}": {
      "get": {
        "tags": [
          "Stations"
        ],
        "summary": "取得單一站點",
        "operationId": "get_station_api_v1_stations__sno__get",
        "parameters": [
          {
            "name": "sno",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sno"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Stations"
        ],
        "summary": "完整更新站點（PUT）",
        "operationId": "replace_station_api_v1_stations__sno__put",
        "parameters": [
          {
            "name": "sno",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sno"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StationBase"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      },
      "patch": {
        "tags": [
          "Stations"
        ],
        "summary": "部分更新站點（PATCH）",
        "operationId": "patch_station_api_v1_stations__sno__patch",
        "parameters": [
          {
            "name": "sno",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sno"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Stations"
        ],
        "summary": "刪除站點",
        "operationId": "delete_station_api_v1_stations__sno__delete",
        "parameters": [
          {
            "name": "sno",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sno"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stations/{sno}/availability": {
      "patch": {
        "tags": [
          "Stations"
        ],
        "summary": "回報即時車況（可借 / 可還）",
        "operationId": "report_availability_api_v1_stations__sno__availability_patch",
        "parameters": [
          {
            "name": "sno",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sno"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilityUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "Message"
      },
      "Page_StationOut_": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "符合條件的總筆數"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/StationOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "Page[StationOut]"
      },
...
      "StationBase": {
        "properties": {
          "name_zh": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name Zh",
            "examples": [
              "YouBike2.0_捷運科技大樓站"
            ]
          },
          "name_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Name En",
            "examples": [
              "YouBike2.0_MRT Technology Bldg. Sta."
            ]
          },
          "area_id": {
            "type": "integer",
            "title": "Area Id",
            "description": "所屬行政區 id（參考 /areas）",
            "examples": [
              1
            ]
          },
          "address_zh": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Zh",
            "examples": [
              "復興南路二段235號前"
            ]
          },
          "address_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Address En",
            "examples": [
              "No.235， Sec. 2， Fuxing S. Rd."
            ]
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "examples": [
              25.02605
            ]
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "examples": [
              121.5436
            ]
          },
          "total_docks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Docks",
            "description": "總車位數",
            "examples": [
              28
            ]
          },
          "available_bikes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Available Bikes",
            "description": "可借車輛數",
            "examples": [
              5
            ]
          },
          "available_docks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Available Docks",
            "description": "可還空位數",
            "examples": [
              22
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "站點是否營運中",
            "default": true
          },
          "info_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Info Time",
            "description": "站點資料回報時間"
          }
        },
        "type": "object",
        "required": [
          "name_zh",
          "area_id",
          "latitude",
          "longitude",
          "total_docks",
          "available_bikes",
          "available_docks"
        ],
        "title": "StationBase"
      },
      "StationCreate": {
        "properties": {
          "name_zh": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name Zh",
            "examples": [
              "YouBike2.0_捷運科技大樓站"
            ]
          },
          "name_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Name En",
            "examples": [
              "YouBike2.0_MRT Technology Bldg. Sta."
            ]
          },
          "area_id": {
            "type": "integer",
            "title": "Area Id",
            "description": "所屬行政區 id（參考 /areas）",
            "examples": [
              1
            ]
          },
          "address_zh": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Zh",
            "examples": [
              "復興南路二段235號前"
            ]
          },
          "address_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Address En",
            "examples": [
              "No.235， Sec. 2， Fuxing S. Rd."
            ]
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "examples": [
              25.02605
            ]
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "examples": [
              121.5436
            ]
          },
          "total_docks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Docks",
            "description": "總車位數",
            "examples": [
              28
            ]
          },
          "available_bikes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Available Bikes",
            "description": "可借車輛數",
            "examples": [
              5
            ]
          },
          "available_docks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Available Docks",
            "description": "可還空位數",
            "examples": [
              22
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "站點是否營運中",
            "default": true
          },
          "info_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Info Time",
            "description": "站點資料回報時間"
          },
          "sno": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Sno",
            "description": "站點代號（主鍵）",
            "examples": [
              "500199999"
            ]
          }
        },
        "type": "object",
        "required": [
          "name_zh",
          "area_id",
          "latitude",
          "longitude",
          "total_docks",
          "available_bikes",
          "available_docks",
          "sno"
        ],
        "title": "StationCreate"
      },
...
      "StationOut": {
        "properties": {
          "name_zh": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name Zh",
            "examples": [
              "YouBike2.0_捷運科技大樓站"
            ]
          },
          "name_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Name En",
            "examples": [
              "YouBike2.0_MRT Technology Bldg. Sta."
            ]
          },
          "area_id": {
            "type": "integer",
            "title": "Area Id",
            "description": "所屬行政區 id（參考 /areas）",
            "examples": [
              1
            ]
          },
          "address_zh": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Zh",
            "examples": [
              "復興南路二段235號前"
            ]
          },
          "address_en": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Address En",
            "examples": [
              "No.235， Sec. 2， Fuxing S. Rd."
            ]
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "examples": [
              25.02605
            ]
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "examples": [
              121.5436
            ]
          },
          "total_docks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Docks",
            "description": "總車位數",
            "examples": [
              28
            ]
          },
          "available_bikes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Available Bikes",
            "description": "可借車輛數",
            "examples": [
              5
            ]
          },
          "available_docks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Available Docks",
            "description": "可還空位數",
            "examples": [
              22
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "站點是否營運中",
            "default": true
          },
          "info_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Info Time",
            "description": "站點資料回報時間"
          },
          "sno": {
            "type": "string",
            "title": "Sno"
          },
          "area_name": {
            "type": "string",
            "title": "Area Name",
            "description": "行政區中文名稱"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name_zh",
          "area_id",
          "latitude",
          "longitude",
          "total_docks",
          "available_bikes",
          "available_docks",
          "sno",
          "area_name",
          "updated_at"
        ],
        "title": "StationOut"
      },
...
    }
  },
...
}